# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hotel competitor monitor API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/watchlist": {
   "get": {
    "tags": [
     "watchlist"
    ],
    "summary": "List Watchlist",
    "operationId": "list_watchlist_watchlist_get",
    "parameters": [
     {
      "name": "include_inactive",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Inactive"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/WatchItemOut"
         },
         "title": "Response List Watchlist Watchlist Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "watchlist"
    ],
    "summary": "Add Hotel",
    "description": "Thêm khách sạn bằng URL Booking.com. Listing đã có trong hệ thống (tenant khác theo dõi) thì\ndùng chung khách sạn đó; listing mới được tạo `unverified` và đẩy job kiểm tra (tên, toạ độ).",
    "operationId": "add_hotel_watchlist_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WatchItemCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WatchItemOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/compset-review": {
   "get": {
    "tags": [
     "watchlist"
    ],
    "summary": "Compset Review",
    "description": "Rà soát compset của một khách sạn của bạn theo quy tắc CoStar STR (7.3).",
    "operationId": "compset_review_watchlist_compset_review_get",
    "parameters": [
     {
      "name": "own_hotel_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Own Hotel Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompsetReviewOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/{hotel_id}": {
   "patch": {
    "tags": [
     "watchlist"
    ],
    "summary": "Update Item",
    "operationId": "update_item_watchlist__hotel_id__patch",
    "parameters": [
     {
      "name": "hotel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hotel Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WatchItemUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WatchItemOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "watchlist"
    ],
    "summary": "Remove Item",
    "operationId": "remove_item_watchlist__hotel_id__delete",
    "parameters": [
     {
      "name": "hotel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hotel Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/{hotel_id}/listings": {
   "post": {
    "tags": [
     "watchlist"
    ],
    "summary": "Add Listing",
    "description": "Sửa URL Booking.com của khách sạn (listing hỏng hoặc dán nhầm). Thay URL đang quét chỉ khi\ntenant là người theo dõi duy nhất.",
    "operationId": "add_listing_watchlist__hotel_id__listings_post",
    "parameters": [
     {
      "name": "hotel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hotel Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ListingOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/{hotel_id}/listings/{listing_id}": {
   "patch": {
    "tags": [
     "watchlist"
    ],
    "summary": "Act On Listing",
    "operationId": "act_on_listing_watchlist__hotel_id__listings__listing_id__patch",
    "parameters": [
     {
      "name": "hotel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hotel Id"
      }
     },
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingAction"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/ListingOut"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Act On Listing Watchlist  Hotel Id  Listings  Listing Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/scan-now": {
   "post": {
    "tags": [
     "watchlist"
    ],
    "summary": "Scan Now",
    "description": "Quét ngay toàn bộ watchlist của tenant (mỗi kênh một run), không đợi mốc giờ.",
    "operationId": "scan_now_watchlist_scan_now_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ScanRunOut"
         },
         "title": "Response Scan Now Watchlist Scan Now Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/watchlist/{hotel_id}/scan-now": {
   "post": {
    "tags": [
     "watchlist"
    ],
    "summary": "Scan Hotel Now",
    "description": "Quét ngay một khách sạn trong watchlist (mỗi kênh đang quét một run nhỏ).",
    "operationId": "scan_hotel_now_watchlist__hotel_id__scan_now_post",
    "parameters": [
     {
      "name": "hotel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hotel Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ScanRunOut"
         },
         "title": "Response Scan Hotel Now Watchlist  Hotel Id  Scan Now Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CompsetReviewOut": {
    "properties": {
     "own_hotel_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Own Hotel Id"
     },
     "primary": {
      "type": "integer",
      "title": "Primary"
     },
     "secondary": {
      "type": "integer",
      "title": "Secondary"
     },
     "rooms_known": {
      "type": "integer",
      "title": "Rooms Known"
     },
     "warnings": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Warnings"
     },
     "dominant_hotel_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dominant Hotel Id"
     },
     "dominant_share": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dominant Share"
     },
     "last_change_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Change At"
     }
    },
    "type": "object",
    "required": [
     "own_hotel_id",
     "primary",
     "secondary",
     "rooms_known",
     "warnings"
    ],
    "title": "CompsetReviewOut",
    "description": "Rà soát compset theo quy tắc CoStar STR (7.3): ≥4 đối thủ, ≥3 không cùng chủ (chưa kiểm\nđược), không khách sạn nào quá 50% số phòng compset; nhắc rà soát ≥2 lần/năm."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ListingAction": {
    "properties": {
     "action": {
      "type": "string",
      "pattern": "^(pause|resume|retry)$",
      "title": "Action"
     }
    },
    "type": "object",
    "required": [
     "action"
    ],
    "title": "ListingAction"
   },
   "ListingCreate": {
    "properties": {
     "url": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 8,
      "title": "Url"
     }
    },
    "type": "object",
    "required": [
     "url"
    ],
    "title": "ListingCreate"
   },
   "ListingOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "hotel_id": {
      "type": "integer",
      "title": "Hotel Id"
     },
     "channel": {
      "type": "string",
      "title": "Channel"
     },
     "listing_key": {
      "type": "string",
      "title": "Listing Key"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "external_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "External Id"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "match_score": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Match Score"
     },
     "last_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error"
     },
     "verified_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Verified At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "hotel_id",
     "channel",
     "listing_key",
     "url",
     "external_id",
     "name",
     "status",
     "match_score",
     "last_error",
     "verified_at"
    ],
    "title": "ListingOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WatchItemCreate": {
    "properties": {
     "url": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 8,
      "title": "Url"
     },
     "role": {
      "type": "string",
      "pattern": "^(self|competitor)$",
      "title": "Role",
      "default": "competitor"
     },
     "label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     }
    },
    "type": "object",
    "required": [
     "url"
    ],
    "title": "WatchItemCreate"
   },
   "WatchItemOut": {
    "properties": {
     "hotel": {
      "$ref": "#/components/schemas/HotelOut"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "active": {
      "type": "boolean",
      "title": "Active"
     },
     "added_at": {
      "type": "string",
      "format": "date-time",
      "title": "Added At"
     },
     "compset_of": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compset Of"
     },
     "tier": {
      "type": "string",
      "enum": [
       "primary",
       "secondary"
      ],
      "title": "Tier",
      "default": "primary"
     },
     "weight": {
      "type": "string",
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Weight",
      "default": "1"
     }
    },
    "type": "object",
    "required": [
     "hotel",
     "role",
     "label",
     "active",
     "added_at"
    ],
    "title": "WatchItemOut"
   },
   "WatchItemUpdate": {
    "properties": {
     "role": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(self|competitor)$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     },
     "label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Active"
     },
     "compset_of": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compset Of"
     },
     "tier": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "primary",
         "secondary"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier"
     },
     "weight": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10.0,
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weight"
     },
     "rooms_total": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 5000.0,
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Rooms Total"
     }
    },
    "type": "object",
    "title": "WatchItemUpdate"
   },
...
  }
 }
}